chore: CLean up for PHPMD
